alert/requirements.txt
flask
flask_sqlalchemy
flask_login
gunicorn
google-generativeai
python-dotenv
werkzeug